requirements.txt
appnope==0.1.3
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cookiecutter==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
decorator==5.1.1
Deprecated==1.2.13
distlib==0.3.7
dj-rest-auth==3.0.0
Django==4.2.1
django-autoslug==1.9.9
django-browser-reload==1.8.0
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
django-extensions==3.2.1
django-htmlmin==0.11.0
django-oauth-toolkit==2.2.0
django-polymorphic==3.1.0
django-rest-framework==0.1.0
django-rest-polymorphic==0.1.10
django-split-settings==1.2.0
django-tailwind==3.5.0
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
djangorestframework-csv==2.1.1
djangorestframework-dataclasses==1.2.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
executing==1.2.0
filelock==3.12.2
html5lib==1.1
idna==3.4
inflection==0.5.1
ipython==8.12.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jinja2-time==0.2.0
jwcrypto==1.4.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
oauthlib==3.2.2
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
requests==2.28.2
rest-framework-simplejwt==0.0.2
ruamel.yaml==0.17.21
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
text-unidecode==1.3
traitlets==5.9.0
tzdata==2023.3
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.24.2
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.3
wrapt==1.15.0
